Move twitstream.twitstream contents into __init__.py: import is now back to
